% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{addCohortFolder}
\alias{addCohortFolder}
\title{Function to create a cohort folder in input/cohortsToCreate}
\usage{
addCohortFolder(folderName, projectPath = here::here())
}
\arguments{
\item{folderName}{The name of the new folder}

\item{projectPath}{the path to the project}
}
\description{
Function to create a cohort folder in input/cohortsToCreate
}
