% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{addConfig}
\alias{addConfig}
\title{Add a line to the config file}
\usage{
addConfig(block, database = block, projectPath = here::here(), open = TRUE)
}
\arguments{
\item{block}{the name of the config block}

\item{database}{the name of the database for the block, default to block name}

\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\description{
Add a line to the config file
}
