% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{addDataSources}
\alias{addDataSources}
\title{Function to add study data sources}
\usage{
addDataSources(..., projectPath = here::here())
}
\arguments{
\item{...}{a list of data sources to add}

\item{projectPath}{the path to the project}
}
\value{
invisible studyYml, prints specifying what changes happened
}
\description{
Function to add study data sources
}
