% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{addGitRemoteToUlysses}
\alias{addGitRemoteToUlysses}
\title{Function to add a Remote to Ulysses directory}
\usage{
addGitRemoteToUlysses(
  gitRemoteUrl,
  gitRemoteName = "origin",
  commitMessage = NULL
)
}
\arguments{
\item{gitRemoteUrl}{a character string of a git remote url}

\item{gitRemoteName}{The name of the remote, defaults to origin}

\item{commitMessage}{a character string of a commit Message to use. if null then skips commit}
}
\description{
This function adds a git remote to the Ulysses repo. If user adds a commit message
it will add and commit files prior to adding and pushing to remote. This function
will check to see if there are untracked files that need to be commited prior to adding remote.
}
