% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{addLinks}
\alias{addLinks}
\title{Function to add study links}
\usage{
addLinks(..., projectPath = here::here())
}
\arguments{
\item{...}{a list of links to add}

\item{projectPath}{the path to the project}
}
\value{
invisible studyYml, prints specifying what changes happened
}
\description{
Function to add study links
}
