% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{addScratchFolder}
\alias{addScratchFolder}
\title{Function to create a scratch folder}
\usage{
addScratchFolder(projectPath = here::here())
}
\arguments{
\item{projectPath}{the path to the project}
}
\description{
Function to create a scratch folder
}
