% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{addStudyMember}
\alias{addStudyMember}
\title{Function to add a study member}
\usage{
addStudyMember(name, email = NULL, projectPath = here::here())
}
\arguments{
\item{name}{the name of the study member}

\item{email}{the study member's email}

\item{projectPath}{the path to the project}
}
\value{
invisible studyYml, prints specifying what changes happened
}
\description{
Function to add a study member
}
