% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{addTags}
\alias{addTags}
\title{Function to add study tags}
\usage{
addTags(..., projectPath = here::here())
}
\arguments{
\item{...}{a list of study tags to add}

\item{projectPath}{the path to the project}
}
\value{
invisible studyYml, prints specifying what changes happened
}
\description{
Function to add study tags
}
