% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyHub.R
\name{buildStudyHub}
\alias{buildStudyHub}
\title{Build Study Hub}
\usage{
buildStudyHub(projectPath = here::here(), previewHub = TRUE)
}
\arguments{
\item{projectPath}{the path to the Ulysses repo, by default takes the path of the active R project}

\item{previewHub}{toggle to preview the hub after it builds. Default is TRUE}
}
\value{
invisible return. Creates _site folder with html files to preview site
}
\description{
Build Study Hub
}
