% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{cloneUlysses}
\alias{cloneUlysses}
\title{Function to clone a Ulysses Repo into local}
\usage{
cloneUlysses(gitRemoteUrl, repoFolder)
}
\arguments{
\item{gitRemoteUrl}{the url of the git remote used to clone}

\item{repoFolder}{the location where you wish to place the git repo}
}
\value{
invisible return. Clones the git remote into local file structure and infuses default folders in clone
}
\description{
Function to clone a Ulysses Repo into local
}
