% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{cohortManifest}
\alias{cohortManifest}
\title{Function that lists all cohort definitions loaded into the study}
\usage{
cohortManifest(projectPath = here::here())
}
\arguments{
\item{projectPath}{the path to the project}
}
\value{
tibble of the cohorts in the project
}
\description{
Function that lists all cohort definitions loaded into the study
}
