% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{defaultCredentials}
\alias{defaultCredentials}
\title{Function to list default credentials}
\usage{
defaultCredentials()
}
\description{
This function builds the standard credential set needed for most connections. If
another credential is needed use \code{c()} to bind the character vector.
}
