% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{defineLoadTable}
\alias{defineLoadTable}
\title{Define the load table}
\usage{
defineLoadTable(atlasId, label, category, subCategory)
}
\arguments{
\item{label}{a vector of naming labels to identify the atlas Ids, must be in order of atlas ids}

\item{category}{a vector of categories to identify the atlas ids, must be in order of atlas ids}

\item{subCategory}{a vector of sub-categories to identify the atlas ids, must be in order of atlas ids}

\item{atlas}{a vector of atlas Ids to load}
}
\value{
a tibble where each row is an atlas asset with meta data
}
\description{
Define the load table
}
