% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execStudyPipeline}
\alias{execStudyPipeline}
\title{Function to execute all study task in analysis folder on set of configBlock}
\usage{
execStudyPipeline(configBlock, env = rlang::caller_env())
}
\arguments{
\item{configBlock}{name of one or multiple configBlock to use in the execution}

\item{env}{the execution environment}
}
\description{
Function to execute all study task in analysis folder on set of configBlock
}
