% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execStudyTask}
\alias{execStudyTask}
\title{Function to execute a study task in Ulysses}
\usage{
execStudyTask(taskFile, configBlock, env = rlang::caller_env())
}
\arguments{
\item{taskFile}{the name of the taskFile. Only use the base name}

\item{configBlock}{the name of the configBlock to use in the execution}

\item{env}{the execution environment}
}
\description{
Function to execute a study task in Ulysses
}
