% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{ExecutionSettings}
\alias{ExecutionSettings}
\title{ExecutionSettings}
\description{
An R6 class to define an ExecutionSettings object
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{cdmDatabaseSchema}}{the schema containing the OMOP CDM}

\item{\code{workDatabaseSchema}}{the schema containing the cohort table}

\item{\code{tempEmulationSchema}}{the schema needed for temp tables}

\item{\code{cohortTable}}{the table containing the cohorts}

\item{\code{databaseName}}{the name of the source data of the cdm}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExecutionSettings-new}{\code{ExecutionSettings$new()}}
\item \href{#method-ExecutionSettings-getDbms}{\code{ExecutionSettings$getDbms()}}
\item \href{#method-ExecutionSettings-connect}{\code{ExecutionSettings$connect()}}
\item \href{#method-ExecutionSettings-disconnect}{\code{ExecutionSettings$disconnect()}}
\item \href{#method-ExecutionSettings-getConnection}{\code{ExecutionSettings$getConnection()}}
\item \href{#method-ExecutionSettings-clone}{\code{ExecutionSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionSettings-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionSettings$new(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema = NULL,
  workDatabaseSchema = NULL,
  tempEmulationSchema = NULL,
  cohortTable = NULL,
  databaseName = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionDetails}}{a connectionDetails object}

\item{\code{connection}}{a connection to a dbms}

\item{\code{cdmDatabaseSchema}}{The schema of the OMOP CDM database}

\item{\code{workDatabaseSchema}}{The schema to which results will be written}

\item{\code{tempEmulationSchema}}{Some database platforms like Oracle and Snowflake do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created.}

\item{\code{cohortTable}}{The name of the table where the cohort(s) are stored}

\item{\code{databaseName}}{A human-readable name for the OMOP CDM database}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionSettings-getDbms"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionSettings-getDbms}{}}}
\subsection{Method \code{getDbms()}}{
extract the dbms dialect
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionSettings$getDbms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionSettings-connect"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionSettings-connect}{}}}
\subsection{Method \code{connect()}}{
connect to dbms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionSettings$connect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionSettings-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionSettings-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
disconnect from dbms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionSettings$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionSettings-getConnection"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionSettings-getConnection}{}}}
\subsection{Method \code{getConnection()}}{
retrieve the connection object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionSettings$getConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
