% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{getAnalysisCohortsByCategory}
\alias{getAnalysisCohortsByCategory}
\title{function to subset the cohortManifest based on category and sub category}
\usage{
getAnalysisCohortsByCategory(
  cohortCategory,
  cohortSubCategory = NULL,
  cohortManifestPath = here::here("inputs/cohorts/cohortManifest.csv")
)
}
\arguments{
\item{cohortCategory}{the category label in the cohort manifest}

\item{cohortSubCategory}{the subCategory label to filter the cohort manifest, default is null}

\item{cohortManifestPath}{the path to the manifest. Defaults to the standard ulysses path}
}
\value{
a tibble subsetting the cohort manifest by the id and label
}
\description{
function to subset the cohortManifest based on category and sub category
}
