% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{getHadesWideLock}
\alias{getHadesWideLock}
\title{Function to grab the hades wide lock file}
\usage{
getHadesWideLock(version, projectPath = here::here())
}
\arguments{
\item{version}{the version of the hades wide lock file to use}

\item{projectPath}{the path to the project}
}
\description{
Function to grab the hades wide lock file
}
