% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R
\name{importAtlasCohorts}
\alias{importAtlasCohorts}
\title{Function to import Atlas cohorts into Ulysses}
\usage{
importAtlasCohorts(cohortIds, keyringName = "atlas", keyringPassword = "ohdsi")
}
\arguments{
\item{cohortIds}{the atlas ids of the cohorts you want to import}

\item{keyringName}{the name of the keyring to save credentials. Defaults to atlas}

\item{keyringPassword}{the password for the keyring to save credentials. Defaults to ohdsi}
}
\value{
saves circe json of ids given to the cohorts/json folder of Ulysses
}
\description{
Function to import Atlas cohorts into Ulysses
}
