% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{importAtlasCohortsFromManifest}
\alias{importAtlasCohortsFromManifest}
\title{Import Atlas Cohorts from the manifest}
\usage{
importAtlasCohortsFromManifest(
  cohortManifest,
  atlasConnection,
  outputFolder = here::here("inputs/cohorts/json")
)
}
\arguments{
\item{atlasConnection}{an WebApiConnection R6 class that holds the creds to connect to webapi}

\item{outputFolder}{the location where the cohorts should be written}

\item{conceptSetManifest}{the location of the cohort manifest}
}
\value{
invisible return but stores the cohort files into Ulysses
}
\description{
Import Atlas Cohorts from the manifest
}
