% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R
\name{importAtlasConceptSets}
\alias{importAtlasConceptSets}
\title{Function to import Atlas concept sets into Ulysses}
\usage{
importAtlasConceptSets(
  conceptSetIds,
  keyringName = "atlas",
  keyringPassword = "ohdsi",
  savePath = here::here("cohorts/conceptSets/json")
)
}
\arguments{
\item{conceptSetIds}{the atlas ids of the cconcept sets you want to import}

\item{keyringName}{the name of the keyring to save credentials. Defaults to atlas}

\item{keyringPassword}{the password for the keyring to save credentials. Defaults to ohdsi}

\item{savePath}{the location in the repo to save the concept set json files to. Defaults to cohorts/conceptSets/json}
}
\value{
saves circe json of ids given to the cohorts/conceptSets/json folder of Ulysses
}
\description{
Function to import Atlas concept sets into Ulysses
}
