% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{importAtlasConceptSetsFromManifest}
\alias{importAtlasConceptSetsFromManifest}
\title{Import Atlas Concept sets from the manifest}
\usage{
importAtlasConceptSetsFromManifest(
  conceptSetManifest,
  atlasConnection,
  outputFolder = here::here("inputs/conceptSets/json")
)
}
\arguments{
\item{conceptSetManifest}{the location of the concept set manifest}

\item{atlasConnection}{an WebApiConnection R6 class that holds the creds to connect to webapi}

\item{outputFolder}{the location where the concept sets should be written}
}
\value{
invisible return but stores the concept set files into Ulysses
}
\description{
Import Atlas Concept sets from the manifest
}
