% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{importCredentialsToConfig}
\alias{importCredentialsToConfig}
\title{Function to import credentials in stored csv to study config.yml}
\usage{
importCredentialsToConfig(credFile, projectPath = here::here(), open = TRUE)
}
\arguments{
\item{credFile}{a credential.csv file stored in a secure location}

\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\description{
Function to import credentials in stored csv to study config.yml
}
