% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{importImages}
\alias{importImages}
\title{Function to import a folder of images for a study hub}
\usage{
importImages(imageFolder, projectPath = here::here())
}
\arguments{
\item{imageFolder}{the file path for an image folder}

\item{projectPath}{path to ohdsi study}
}
\value{
invisible return of the image file folder
}
\description{
Function to import a folder of images for a study hub
}
