% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{initConfigUsingKeyring}
\alias{initConfigUsingKeyring}
\title{Function to initialize config.yml using keyring}
\usage{
initConfigUsingKeyring(dbIds, projectPath = here::here(), open = TRUE)
}
\arguments{
\item{dbIds}{the dbIds to add to the config file}

\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\description{
Function to initialize config.yml using keyring
}
