% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{initConfigUsingTable}
\alias{initConfigUsingTable}
\title{Function to import credentials in stored csv to study config.yml}
\usage{
initConfigUsingTable(
  credFile = fs::path_home("shhh.csv"),
  projectPath = here::here(),
  open = TRUE
)
}
\arguments{
\item{credFile}{a credential.csv file stored in a secure location}

\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\description{
Function to import credentials in stored csv to study config.yml
}
