% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{initCredentialsKeyring}
\alias{initCredentialsKeyring}
\title{Function to set multi credentials}
\usage{
initCredentialsKeyring(db_id, keyringPassword = "ulysses", forceCheck = TRUE)
}
\arguments{
\item{db_id}{the database id to distinguish the set of credentials}

\item{keyringPassword}{the password for the keyring. Defaults to ulysses.}

\item{forceCheck}{a toggle that will print blurred credentials to check credential}
}
\description{
Function to set multi credentials
}
