% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsiStudy.R
\name{isOhdsiStudy}
\alias{isOhdsiStudy}
\title{Function to check if the directory is a picard project}
\usage{
isOhdsiStudy(basePath)
}
\arguments{
\item{basePath}{the path of the directory}
}
\description{
Function to check if the directory is a picard project
}
