% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{launchUlyssesRemoteWithBitBucketDC}
\alias{launchUlyssesRemoteWithBitBucketDC}
\title{Launch Ulysses Remote using Bitbucket Data Center}
\usage{
launchUlyssesRemoteWithBitBucketDC(repoName, hostUrl, httpToken, projectName)
}
\arguments{
\item{repoName}{the name of the repository in project}

\item{hostUrl}{the url hosting bitbucket data center}

\item{httpToken}{the http access token configured to your bitbucket profile. To find the httpToken go to Manage Account > HTTP access tokens.
It is recommended that you store the httpToken as an environment variable fo ease of use.}

\item{projectName}{the name of the project in Bitbucket Data Center storing the repo}
}
\value{
invisible return but initializes the remote on BitbucketDC
}
\description{
Launch Ulysses Remote using Bitbucket Data Center
}
