% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClasses.R
\name{launchUlyssesRepo}
\alias{launchUlyssesRepo}
\title{Function to Launch new Ulysses Repo}
\usage{
launchUlyssesRepo(ulyssesStudySettings, verbose = TRUE, openProject = FALSE)
}
\arguments{
\item{ulyssesStudySettings}{UlyssesStudy R6 class with the ulysses study details to make}

\item{verbose}{a toggle whether to print details of launch in console}

\item{openProject}{a toggle whether to open the repo as a new R project}
}
\value{
invisible return. Creates the ulysses repo in the local file structure
}
\description{
Function to Launch new Ulysses Repo
}
