% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{makeCaprScript}
\alias{makeCaprScript}
\title{Function to create a pipeline task as an R file}
\usage{
makeCaprScript(
  scriptName,
  configBlock = NULL,
  projectPath = here::here(),
  open = TRUE
)
}
\arguments{
\item{scriptName}{The name of the capr file that is being created}

\item{configBlock}{the name of the config block to use for the script}

\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\description{
Function to create a pipeline task as an R file
}
