% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{makeCohortDetails}
\alias{makeCohortDetails}
\title{Function to create the cohort details for the study}
\usage{
makeCohortDetails(projectPath = here::here(), open = TRUE)
}
\arguments{
\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\value{
writes a CohortDetails.qmd file to the documentation folder
}
\description{
The cohort details file is a document that provides a human-readable version
of the circe cohort definitions used in the study. This function targets
the cohortsToCreate folder and makes a quarto of the cohort details.
}
