% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{makeContributionGuidelines}
\alias{makeContributionGuidelines}
\title{R Markdown file to make the contribution guidelines}
\usage{
makeContributionGuidelines(projectPath = here::here(), open = TRUE)
}
\arguments{
\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\description{
R Markdown file to make the contribution guidelines
}
