% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{makeExample}
\alias{makeExample}
\title{Function to create an example OHDSI script}
\usage{
makeExample(fileName, savePath = here::here(), open = TRUE)
}
\arguments{
\item{fileName}{The name of for the file}

\item{savePath}{the path to save the file, defaults to active directory}

\item{open}{toggle on whether the file should be opened}
}
\description{
Function to create an example OHDSI script
}
