% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClasses.R
\name{makeExecOptions}
\alias{makeExecOptions}
\title{Make ExecOptions for Ulysses}
\usage{
makeExecOptions(
  dbms,
  workDatabaseSchema,
  tempEmulationSchema = NULL,
  dbConnectionBlocks
)
}
\arguments{
\item{dbms}{specify the dbms used in the exec options}

\item{workDatabaseSchema}{the name of the workDatabaseSchema as a character string, location in DB where user has write access}

\item{tempEmulationSchema}{he name of the tempEmulationSchema as a character strings}

\item{dbConnectionBlocks}{a list of DbConfigBlock R6 classes specifying the dbs to connect}
}
\value{
A ExecOptions R6 class with the execOptions
}
\description{
Make ExecOptions for Ulysses
}
