% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{makeHowToRun}
\alias{makeHowToRun}
\title{Function to create a HowToRun file}
\usage{
makeHowToRun(projectPath = here::here(), open = TRUE)
}
\arguments{
\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\description{
Function to create a HowToRun file
}
