% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{makeInternals}
\alias{makeInternals}
\title{Function to create a pipeline task as an R file}
\usage{
makeInternals(
  internalsName,
  projectPath = here::here(),
  author = NULL,
  date = lubridate::today(),
  open = TRUE
)
}
\arguments{
\item{internalsName}{The name of the internals file that is being created}

\item{projectPath}{the path to the project}

\item{author}{the author of the R file, defaults to first author in _study.yml}

\item{date}{the date the file was initialized, defaults to today}

\item{open}{toggle on whether the file should be opened}
}
\description{
Function to create a pipeline task as an R file
}
