% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{makeMeetingMinutes}
\alias{makeMeetingMinutes}
\title{Function to create a meeting minutes file}
\usage{
makeMeetingMinutes(projectPath = here::here(), open = TRUE)
}
\arguments{
\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\description{
Function to create a meeting minutes file
}
