% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{makeMigrationScript}
\alias{makeMigrationScript}
\title{Function to create a migration script}
\usage{
makeMigrationScript(
  scriptName,
  date = lubridate::today(),
  projectPath = here::here(),
  open = TRUE
)
}
\arguments{
\item{scriptName}{The name of the analysis script}

\item{date}{the date the script was built, default to today's date}

\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\description{
Function to create a migration script
}
