% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{makeOhdsiProtocol}
\alias{makeOhdsiProtocol}
\title{R Markdown file to make the ohdsi protocol}
\usage{
makeOhdsiProtocol(projectPath = here::here(), open = TRUE)
}
\arguments{
\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\description{
R Markdown file to make the ohdsi protocol
}
