% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{makeResultsReport}
\alias{makeResultsReport}
\title{Quarto file to make a results report}
\usage{
makeResultsReport(projectPath = here::here(), open = TRUE)
}
\arguments{
\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\description{
Quarto file to make a results report
}
