% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{makeTaskFile}
\alias{makeTaskFile}
\title{Function initializing an R file for an analysis task}
\usage{
makeTaskFile(
  nameOfTask,
  author = NULL,
  description = NULL,
  projectPath = here::here(),
  openFile = TRUE
)
}
\arguments{
\item{nameOfTask}{The name of the analysis task script}

\item{author}{the name of the person authoring the file. Defaults to template text if NULL}

\item{description}{a description of the analysis task. Defaults to template text if NULL}

\item{projectPath}{the path to the project}

\item{openFile}{toggle on whether the file should be opened}
}
\description{
Function initializing an R file for an analysis task
}
