% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{makeWebApiScript}
\alias{makeWebApiScript}
\title{Function to create a pipeline task as an R file}
\usage{
makeWebApiScript(
  keyringName = NULL,
  keyringPassword = NULL,
  projectPath = here::here(),
  open = TRUE
)
}
\arguments{
\item{keyringName}{The name of the keyring for the study, if null defaults to dir basename}

\item{keyringPassword}{The password for the keyring, if null defauls to ohdsi}

\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\description{
Function to create a pipeline task as an R file
}
