% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategus.R
\name{moduleTable}
\alias{moduleTable}
\title{Function to retrieve table of modules}
\usage{
moduleTable(projectPath = here::here())
}
\arguments{
\item{projectPath}{the path to the project}
}
\description{
Function to retrieve table of modules
}
