% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openCredentials}
\alias{openCredentials}
\title{Function to open the credentials table to edit or peak}
\usage{
openCredentials(credsPath = fs::path_home("shhh.csv"))
}
\arguments{
\item{credsPath}{the path to shhh.csv}
}
\value{
tibble of credentials
}
\description{
Function to open the credentials table to edit or peak
}
