% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClasses.R
\name{placeHolderExecOptions}
\alias{placeHolderExecOptions}
\title{set the execOptions as placeholder.}
\usage{
placeHolderExecOptions()
}
\value{
A ExecOptions R6 class with the execOptions
}
\description{
use this function if there is not a dbms connection for the study such as using a proprietary tool for the analysis
}
