% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{populateManifest}
\alias{populateManifest}
\title{Initialize Manifests}
\usage{
populateManifest(manifestType, importFromAtlas = TRUE)
}
\arguments{
\item{manifestType}{the type of manifest to initialize, either conceptSet or cohort}

\item{importFromAtlas}{toggle whether to import content from atlas. Default is TRUE}
}
\value{
invisible return but populates the manifest in its appropriate folder location
}
\description{
Initialize Manifests
}
