% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{previewStudyHub}
\alias{previewStudyHub}
\title{Function to preview study hub}
\usage{
previewStudyHub(projectPath = here::here())
}
\arguments{
\item{projectPath}{path to ohdsi study}
}
\value{
previews study hub in browser
}
\description{
Function to preview study hub
}
