% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{publishStudyToRepository}
\alias{publishStudyToRepository}
\title{Publish Study to Repo}
\usage{
publishStudyToRepository(repoUrl, message = "Initial commit for OHDSI study")
}
\arguments{
\item{repoUrl}{an https url for the repo remote. Example: https://github.com/Org/repo.git}

\item{message}{the commit message for your OHDSI study}
}
\description{
Publish Study to Repo
}
