% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{requestStudyParticipation}
\alias{requestStudyParticipation}
\title{Email asking to participant in an ohdsi study}
\usage{
requestStudyParticipation(
  senderName,
  projectPath = here::here(),
  recipientName = NULL,
  open = TRUE
)
}
\arguments{
\item{senderName}{the name of the person sending the email}

\item{projectPath}{the path to the picard project}

\item{recipientName}{the name of the person receiving the email}

\item{open}{toggle on whether the file should be opened}
}
\description{
Email asking to participant in an ohdsi study
}
