% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{requestStudyRepo}
\alias{requestStudyRepo}
\title{Email asking to initialize an ohdsi-studies repo}
\usage{
requestStudyRepo(
  senderName,
  projectPath = here::here(),
  recipientName = NULL,
  open = TRUE
)
}
\arguments{
\item{senderName}{the name of the person sending the email}

\item{projectPath}{the path to the picard project}

\item{recipientName}{the name of the person receiving the email}

\item{open}{toggle on whether the file should be opened}
}
\description{
Email asking to initialize an ohdsi-studies repo
}
