% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{retrieveStudySettings}
\alias{retrieveStudySettings}
\title{Function to convert yml into a specified list format}
\usage{
retrieveStudySettings(projectPath)
}
\arguments{
\item{projectPath}{the path to the project}
}
\value{
the study yml as an R object (list)
}
\description{
Function to convert yml into a specified list format
}
