% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyHub.R
\name{rsConnectStudyHub}
\alias{rsConnectStudyHub}
\title{Function to publish study hub via rsConnect}
\usage{
rsConnectStudyHub(
  inputPath = here::here("dissemination/quarto"),
  server,
  account,
  appName,
  appTitle
)
}
\arguments{
\item{inputPath}{path to the quarto folder to publish defaults to dissemination/quarto}

\item{server}{the server url}

\item{account}{the account affiliated with the server}

\item{appName}{the name of the app}

\item{appTitle}{the title for the app}
}
\description{
Function to publish study hub via rsConnect
}
