% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{runTaskInBackground}
\alias{runTaskInBackground}
\title{Function to run study task in background}
\usage{
runTaskInBackground(taskFile, configBlock, projectPath = here::here())
}
\arguments{
\item{taskFile}{a task file to run in the background}

\item{configBlock}{the configBlock to use in the study}

\item{projectPath}{the path to the project}
}
\value{
invisble return, launches background job
}
\description{
Function to run study task in background
}
