% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{setAllDatabaseCredentials}
\alias{setAllDatabaseCredentials}
\title{Function to set all database credentials}
\usage{
setAllDatabaseCredentials(cred, db, study, forceCheck = TRUE)
}
\arguments{
\item{cred}{a vector of credentials to set (i.e dbms, user, connectionString). See defaultCredentials on building set}

\item{db}{the database prefix for the credential}

\item{study}{the name of the study for the credential set, this will be the keyring name}

\item{forceCheck}{a toggle that will print blurred credentials to check credential}
}
\description{
Function to set all database credentials
}
