% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R
\name{setAtlasCredentials}
\alias{setAtlasCredentials}
\title{Function to set atlas credentials for keyring}
\usage{
setAtlasCredentials(keyringName = "atlas", keyringPassword = "ohdsi")
}
\arguments{
\item{keyringName}{the name of the keyring to save credentials. Defaults to atlas}

\item{keyringPassword}{the password for the keyring to save credentials. Defaults to ohdsi}
}
\value{
a series of dialog box inputs for the webapi credentials including the baseUrl,
authMethod, user and password
}
\description{
Function to set atlas credentials for keyring
}
