% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClasses.R
\name{setContributor}
\alias{setContributor}
\title{Set Ulysses Contributor}
\usage{
setContributor(name, email, role)
}
\arguments{
\item{name}{the name of the contributor as a character string}

\item{email}{the email of the contributor as a character string}

\item{role}{the role of the contirbutor as a character string}
}
\value{
A ContributorLine R6 class with the contributor info
}
\description{
Set Ulysses Contributor
}
