% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{setCredential}
\alias{setCredential}
\title{Function to set single credential}
\usage{
setCredential(cred, db, keyringName, keyringPassword, forceCheck = TRUE)
}
\arguments{
\item{cred}{the credential to set (i.e dbms, user, connectionString)}

\item{db}{the database prefix for the credential}

\item{keyringName}{the name of the keyringName for the credential set, this will be the keyring namec}

\item{keyringPassword}{the password for the keyring.}

\item{forceCheck}{a toggle that will print blurred credentials to check credential}
}
\description{
Function to set single credential
}
