% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClasses.R
\name{setDbConfigBlock}
\alias{setDbConfigBlock}
\title{set the config block for a database}
\usage{
setDbConfigBlock(
  configBlockName,
  cdmDatabaseSchema,
  databaseName = NULL,
  databaseLabel = NULL
)
}
\arguments{
\item{configBlockName}{the name of the config block}

\item{cdmDatabaseSchema}{the cdmDatabaseSchema specified as a character string}

\item{databaseName}{the name of the database, typically uses the db name and id. For example optum_dod_202501}

\item{databaseLabel}{the labelling name of the database, typically a common name for a db. For example Optum DOD}
}
\value{
A StudyMeta R6 class with the study meta
}
\description{
set the config block for a database
}
