% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{setMultipleCredentials}
\alias{setMultipleCredentials}
\title{Function to set multi credentials}
\usage{
setMultipleCredentials(
  creds,
  db,
  keyringName,
  keyringPassword,
  forceCheck = TRUE
)
}
\arguments{
\item{creds}{a vector of credentials to set (i.e dbms, user, connectionString). See defaultCredentials on building set}

\item{db}{the database prefix for the credential}

\item{keyringName}{the name of the keyringName where the credential will be set}

\item{forceCheck}{a toggle that will print blurred credentials to check credential}

\item{keyringPasssword}{the password for the keyring.}
}
\description{
Function to set multi credentials
}
