% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategus.R
\name{setStrategusInstantiatedFolder}
\alias{setStrategusInstantiatedFolder}
\title{Function to set strategus instantiated folder}
\usage{
setStrategusInstantiatedFolder(folderName, root = NULL)
}
\arguments{
\item{folderName}{the name of the instantiated folder}

\item{root}{the root path to the folder, defaults to the home path}
}
\description{
Function to set strategus instantiated folder
}
