% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{setStudyAuthors}
\alias{setStudyAuthors}
\title{Function to set study authors}
\usage{
setStudyAuthors(
  developer = Sys.getenv("USERNAME"),
  developerEmail = glue::glue("{developer}@ohdsi.org"),
  lead = "Ulysses",
  leadEmail = "Ulysses@ohdsi.org"
)
}
\arguments{
\item{developer}{specify the study developer, default to system variable}

\item{developerEmail}{the email of the developer}

\item{lead}{specify the lead of the study}

\item{leadEmail}{specify the lead of the study email}
}
\value{
a list containing study authors
}
\description{
Function to set study authors
}
