% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{setStudyDescription}
\alias{setStudyDescription}
\title{Function to set study description}
\usage{
setStudyDescription(
  desc = "Provide a 1 to 2 sentence description of your study. Be concise.",
  ta = "Specify the therapeutic area of the study",
  dataSources = list("Truven MarketScan", "Optum Market Clarity")
)
}
\arguments{
\item{desc}{a brief description of the study}

\item{ta}{a character string of the therapeutic area}

\item{dataSources}{a list of data sources}
}
\value{
a list containing study description info
}
\description{
Function to set study description
}
