% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{setStudyInfo}
\alias{setStudyInfo}
\title{Function to set study info}
\usage{
setStudyInfo(
  id,
  title = id,
  type = c("Characterization"),
  version = "0.0.0.999"
)
}
\arguments{
\item{id}{specify the study id}

\item{title}{specify the study title}

\item{type}{specify the type of study}

\item{version}{specify the study version}
}
\value{
a list containing study info
}
\description{
Function to set study info
}
