% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{setStudyKeyring}
\alias{setStudyKeyring}
\title{Function to set study keyring}
\usage{
setStudyKeyring(keyringName, keyringPassword)
}
\arguments{
\item{keyringName}{the name of the keyring, this should be the reponame for the study}

\item{keyringPassword}{a password to access the study keyring}
}
\description{
Function to set study keyring
}
