% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{setStudyLinks}
\alias{setStudyLinks}
\title{Function to set study links}
\usage{
setStudyLinks(...)
}
\arguments{
\item{...}{a series of resource links for the study}
}
\value{
a list containing links to resources
}
\description{
Function to set study links
}
