% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{setStudyTags}
\alias{setStudyTags}
\title{Function to set study tags}
\usage{
setStudyTags(...)
}
\arguments{
\item{...}{a series of tags for the study}
}
\value{
a list containing study tags
}
\description{
Function to set study tags
}
