% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{setStudyTimeline}
\alias{setStudyTimeline}
\title{Function to set study timeline}
\usage{
setStudyTimeline(
  status = "Started",
  startDate = as.character(lubridate::today()),
  endDate = as.character(lubridate::today() + (365 * 2))
)
}
\arguments{
\item{status}{the study status, default is started}

\item{startDate}{the start of the study, defaults to todays date}

\item{endDate}{the date specifying the end (or expected end) of the study}
}
\value{
a list containing study timeline
}
\description{
Function to set study timeline
}
