% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{syncUlyssesWork}
\alias{syncUlyssesWork}
\title{Function to sync local work in Ulysses to remote}
\usage{
syncUlyssesWork(
  commitMessage,
  branch = gert::git_branch(),
  gitRemoteName = "origin"
)
}
\arguments{
\item{commitMessage}{the commit message describing the work done since the last commit}

\item{branch}{the name of the branch you want to commit changes. Defaults to the active branch}

\item{gitRemoteName}{the name of the remote. Defaults to origin.}
}
\description{
Function to sync local work in Ulysses to remote
}
