% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ulysses-package.R
\docType{package}
\name{Ulysses-package}
\alias{Ulysses}
\alias{Ulysses-package}
\title{Ulysses: Automate OHDSI Study Setup}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Automates setup of OHDSI study and provides functions to assist on improving organization and communication of a study.
}
\author{
\strong{Maintainer}: Martin Lavallee \email{martin.lavallee@odysseusinc.com}

}
\keyword{internal}
