% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{updateDeveloperInfo}
\alias{updateDeveloperInfo}
\title{Function to update developer Infor}
\usage{
updateDeveloperInfo(newName, newEmail, projectPath = here::here())
}
\arguments{
\item{newName}{change the developer name}

\item{newEmail}{change the developer email}
}
\value{
invisible studyYml, prints specifying what changes happened
}
\description{
Function to update developer Infor
}
