% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{updateLeadInfo}
\alias{updateLeadInfo}
\title{Function to update lead Infor}
\usage{
updateLeadInfo(newName, newEmail, projectPath = here::here())
}
\arguments{
\item{newName}{change the lead name}

\item{newEmail}{change the lead email}
}
\value{
invisible studyYml, prints specifying what changes happened
}
\description{
Function to update lead Infor
}
