% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{updateStudyDescription}
\alias{updateStudyDescription}
\title{Function to update study description}
\usage{
updateStudyDescription(newStudyDescription, projectPath = here::here())
}
\arguments{
\item{projectPath}{the path to the project}

\item{newStudyVersion}{the new study description}
}
\value{
invisible studyYml, prints specifying what changes happened
}
\description{
Function to update study description
}
