% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{updateStudyEndDate}
\alias{updateStudyEndDate}
\title{Function to update study end date}
\usage{
updateStudyEndDate(newEndDate, projectPath = here::here())
}
\arguments{
\item{newEndDate}{the new end date of the study, suggest format YYYY-MM-DD}

\item{projectPath}{the path to the project}
}
\value{
invisible studyYml, prints specifying what changes happened
}
\description{
Function to update study end date
}
