% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{updateStudyStatus}
\alias{updateStudyStatus}
\title{Function to update study status}
\usage{
updateStudyStatus(
  newStudyStatus = c("Started", "In-Progress", "Stopped", "Completed"),
  projectPath = here::here()
)
}
\arguments{
\item{newStudyStatus}{the new status of the study, accepts Started, In-Progress, Stopped, Completed}

\item{projectPath}{the path to the project}
}
\value{
invisible studyYml, prints specifying what changes happened
}
\description{
Function to update study status
}
