% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{updateStudyTitle}
\alias{updateStudyTitle}
\title{Function to update study title}
\usage{
updateStudyTitle(newStudyTitle, projectPath = here::here())
}
\arguments{
\item{newStudyTitle}{the new study title}

\item{projectPath}{the path to the project}
}
\value{
invisible studyYml, prints specifying what changes happened
}
\description{
Function to update study title
}
