% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{updateStudyVersion}
\alias{updateStudyVersion}
\title{Function to update the study version}
\usage{
updateStudyVersion(versionNumber, projectPath = here::here())
}
\arguments{
\item{versionNumber}{the semantive version number to set as the new project version: 1.0.0}

\item{projectPath}{the path of the project, defaults to the directory of the active Ulysses project}
}
\description{
Function to update the study version
}
