% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{updateTherapeuticArea}
\alias{updateTherapeuticArea}
\title{Function to update study therapeutic area}
\usage{
updateTherapeuticArea(newTA, projectPath = here::here())
}
\arguments{
\item{newTA}{the new study therapeutic area}

\item{projectPath}{the path to the project}
}
\value{
invisible studyYml, prints specifying what changes happened
}
\description{
Function to update study therapeutic area
}
