% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{zipAndArchive}
\alias{zipAndArchive}
\title{Zip and Archive results from a study execution}
\usage{
zipAndArchive(input)
}
\arguments{
\item{input}{the type of files to zip and archive. There are three options exportMerge, exportPretty and site. exportMerge is the merged results in long format. The exportPretty are xlsx files with formatted output from the study. The site is the html files of the studyHub}
}
\value{
invisible return. Stores the input as a zip file in the exec/archive folder
}
\description{
Zip and Archive results from a study execution
}
