% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{zipResults}
\alias{zipResults}
\title{Funciton to compress results into a zip file}
\usage{
zipResults(databaseName, projectPath = here::here())
}
\arguments{
\item{databaseName}{the name of the database folder to compress}

\item{projectPath}{the path to the project}
}
\description{
Funciton to compress results into a zip file
}
