% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{buildStudyHub}
\alias{buildStudyHub}
\title{Function to build study hub}
\usage{
buildStudyHub(projectPath = here::here())
}
\arguments{
\item{projectPath}{path to ohdsi study}
}
\value{
builds a _site folder in documenation that holds the html website
}
\description{
Function to build study hub
}
