% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{checkDatabaseCredential}
\alias{checkDatabaseCredential}
\title{Function to check the database credential}
\usage{
checkDatabaseCredential(cred, keyringName, verbose = TRUE)
}
\arguments{
\item{cred}{the credential to set (i.e dbms, user, connectionString)}

\item{keyringName}{the name of the keyringName for the credential check, this will be the keyring namec}

\item{verbose}{toggle option to print console message}
}
\description{
Function to check the database credential
}
