% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{initializeManifest}
\alias{initializeManifest}
\title{Initialize Manifests}
\usage{
initializeManifest(manifestType, loadTable = NULL, overwrite = FALSE)
}
\arguments{
\item{manifestType}{the type of manifest to initialize, either conceptSet or cohort}

\item{loadTable}{a tibble made using defineLoadTable which prespecifies atlas ids to initialize the manifest}

\item{overwrite}{toggle whether to overwrite existing manifests. Default is FALSE}
}
\value{
invisible return but initializes the manifest in its appropriate folder location
}
\description{
Initialize Manifests
}
