% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{makeInternals}
\alias{makeInternals}
\title{Function to create a pipeline task as an R file}
\usage{
makeInternals(internalsName, projectPath = here::here(), open = TRUE)
}
\arguments{
\item{internalsName}{The name of the internals file that is being created}

\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}
}
\description{
Function to create a pipeline task as an R file
}
