% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{makeKeyringSetup}
\alias{makeKeyringSetup}
\title{Function to create a config.yml file}
\usage{
makeKeyringSetup(projectPath = here::here(), open = TRUE, secret = NULL)
}
\arguments{
\item{projectPath}{the path to the project}

\item{open}{toggle on whether the file should be opened}

\item{secret}{a keyword to use as the keyring password to access credentials}
}
\description{
Function to create a config.yml file
}
