% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{makeStudySettings}
\alias{makeStudySettings}
\title{Function to initialize study settings}
\usage{
makeStudySettings(
  title,
  authors = defaultAuthors(),
  milestones = defaultMilestones(),
  desc = defaultDesc(),
  links = defaultLinks()
)
}
\arguments{
\item{title}{the title of the study}

\item{authors}{the author list for the study}

\item{milestones}{list of milestone information including study status and timeframe for study}

\item{desc}{a list of attributes describing the study including the study type and tages}

\item{links}{a list of links to files used in study}
}
\description{
Function to initialize study settings
}
