% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClasses.R
\name{makeUlyssesStudySettings}
\alias{makeUlyssesStudySettings}
\title{Make Ulysses Study Settings}
\usage{
makeUlyssesStudySettings(
  repoName,
  repoFolder,
  studyMeta,
  execOptions,
  gitRemote = NULL,
  renvLock = NULL
)
}
\arguments{
\item{repoName}{the name of repo as a character string}

\item{repoFolder}{the folder path where the repo is stored in local as a character string}

\item{studyMeta}{a StudyMeta R6 class with the details describing the study}

\item{execOptions}{a ExecOptions R6 class with the execution details needed for the study}

\item{gitRemote}{a remote url used to clone and set remote git}

\item{renvLock}{file path to a renvLock file}
}
\value{
A UlyssesStudy R6 class with the ulysses study details to make
}
\description{
Make Ulysses Study Settings
}
