% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsiStudy.R
\name{newOhdsiStudy}
\alias{newOhdsiStudy}
\title{Function that initializes a new ohdsi study project environment}
\usage{
newOhdsiStudy(
  path,
  projectName = basename(path),
  studySettings = makeStudySettings(title = basename(path)),
  verbose = TRUE,
  openProject = TRUE
)
}
\arguments{
\item{path}{the path where the project sits}

\item{projectName}{the name of the project}

\item{studySettings}{a list of study settings}

\item{verbose}{whether the function should provide steps, default TRUE}

\item{openProject}{should the project be opened if created}
}
\description{
Function that initializes a new ohdsi study project environment
}
