% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsiStudy.R
\name{newOhdsiStudy}
\alias{newOhdsiStudy}
\title{Function that initializes a new ohdsi study project environment}
\usage{
newOhdsiStudy(
  projectName,
  author,
  type,
  directory = here::here(),
  openProject = TRUE
)
}
\arguments{
\item{projectName}{the name of the project}

\item{author}{the name of the study lead}

\item{type}{the type of study either Characterization, PLP or PLE}

\item{directory}{the directory to create the project}

\item{openProject}{should the project be opened if created}
}
\description{
Function that initializes a new ohdsi study project environment
}
