% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFiles.R
\name{requestStudyRepository}
\alias{requestStudyRepository}
\title{Email asking to initialize an ohdsi-studies repo}
\usage{
requestStudyRepository(
  senderName,
  senderEmail,
  recipientName = NULL,
  recipientEmail = NULL,
  projectPath = here::here(),
  open = TRUE
)
}
\arguments{
\item{senderName}{your name as the person sending the email}

\item{senderEmail}{your email address}

\item{recipientName}{the recipients name, defaults to Admin}

\item{recipientEmail}{the recipients email, defaults to a dummy email}

\item{projectPath}{the path to the Ulysses project}

\item{open}{toggle on whether the file should be opened}
}
\description{
Email asking to initialize an ohdsi-studies repo
}
