% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{setDatabaseCredential}
\alias{setDatabaseCredential}
\title{Function to set single database credential}
\usage{
setDatabaseCredential(cred, db, study, forceCheck = TRUE)
}
\arguments{
\item{cred}{the credential to set (i.e dbms, user, connectionString)}

\item{db}{the database prefix for the credential}

\item{study}{the name of the study for the credential set, this will be the keyring name}

\item{forceCheck}{a toggle that will print blurred credentials to check credential}
}
\description{
Function to set single database credential
}
