% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studySettings.R
\name{updateStudyVersion}
\alias{updateStudyVersion}
\title{Function to update study version}
\usage{
updateStudyVersion(newStudyVersion, projectPath = here::here())
}
\arguments{
\item{newStudyVersion}{the new study version}

\item{projectPath}{the path to the project}
}
\value{
invisible studyYml, prints specifying what changes happened
}
\description{
Function to update study version
}
